+++ NEWS.md
@@ -1,7 +1,8 @@
-# nanonext 0.10.2.9006 (development)
+# nanonext 0.10.2.9007 (development)
 
 #### New Features
 
+* `next_config()` enables native extensions when sending using mode 'next'. Registers hook functions for custom serialization and unserialization of reference objects (such as those accessed via an external pointer).
 * `.until()` contains revised behaviour for this synchronisation primitive that will be ported to `until()` in a future version. The function now returns FALSE instead of TRUE if the timeout has been reached.
 
 #### Updates